Switch to an architecture that uses a -sys crate.
