+++ .github/workflows/publish.yml
@@ -35,9 +35,22 @@ jobs:
           ${{ runner.os }}-cargo-index-
 
     - name: Build and test
-      run: cargo test --verbose
+      run: cargo test --workspace --verbose
 
-    - name: Publish to crates.io
+    - name: Publish xlsynth-sys to crates.io
       env:
         CARGO_REGISTRY_TOKEN: ${{ secrets.CRATES_IO_TOKEN }}
-      run: cargo publish --token $CARGO_REGISTRY_TOKEN
+      run: |
+        cd xlsynth-sys
+        cargo publish --token $CARGO_REGISTRY_TOKEN
+
+
+    - name: Wait for crates.io to update
+      run: sleep 60
+
+    - name: Publish xlsynth
+      env:
+        CARGO_REGISTRY_TOKEN: ${{ secrets.CRATES_IO_TOKEN }}
+      run: |
+        cd xlsynth
+        cargo publish --token $CARGO_REGISTRY_TOKEN
