+++ Cargo.toml
@@ -1,36 +1,7 @@
-[package]
-name = "xlsynth"
-version = "0.0.19"
-edition = "2018"
-authors = ["Christopher D. Leary <cdleary@gmail.com>"]
-description = "Accelerated Hardware Synthesis (XLS/XLSynth) via Rust"
-license = "Apache-2.0"
-repository = "https://github.com/xlsynth/xlsynth-crate"
-documentation = "https://docs.rs/xlsynth"
-homepage = "https://github.com/xlsynth/xlsynth-crate"
-exclude = [
-    ".github",
-]
-
-[dependencies]
-libc = "0.2"
-
-[build-dependencies]
-curl = "0.4"
-flate2 = "1.0"
-tar = "0.4"
-
 [workspace]
 members = [
-    ".",
+    "xlsynth",
+    "xlsynth-sys",
     "sample_usage",
 ]
-
-[dev-dependencies]
-docmatic = "0.1.2"
-curl = "0.4"
-toml = "0.5"
-serde_json = "1.0"
-semver = "1.0"
-log = "0.4"
-env_logger = "0.11"
+resolver = "2"
