+++ sample_usage/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-xlsynth = { path = "../" }
+xlsynth = { path = "../xlsynth" }
 log = "0.4"
 env_logger = "0.11"
 num_cpus = "1.0"
