+++ xlsynth/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "xlsynth"
+version = "0.0.20"
+edition = "2018"
+authors = ["Christopher D. Leary <cdleary@gmail.com>"]
+description = "Accelerated Hardware Synthesis (XLS/XLSynth) via Rust"
+license = "Apache-2.0"
+repository = "https://github.com/xlsynth/xlsynth-crate"
+documentation = "https://docs.rs/xlsynth"
+homepage = "https://github.com/xlsynth/xlsynth-crate"
+
+[dependencies]
+xlsynth-sys = {path = "../xlsynth-sys"}
+
+[dev-dependencies]
+docmatic = "0.1.2"
+curl = "0.4"
+toml = "0.5"
+serde_json = "1.0"
+semver = "1.0"
+log = "0.4"
+env_logger = "0.11"
