+++ xlsynth/src/xlsynth_error.rs
